Refactored All Player related Classes into new package.
Wanted to declutter the model package since it has become a lot less
organized. The PlayerColor interface has also been refactored to
IPlayerColor to easily show the fact that it is an interface.
