+++ Final_Project/src/main/java/connect4Menu/model/ObservableSquarePlayed.java
@@ -1,5 +1,6 @@
 package connect4Menu.model;
 
+import connect4Menu.model.player.Player;
 import connect4Menu.view.ObserverSquarePlayed;
 
 public interface ObservableSquarePlayed {
