+++ Final_Project/src/main/java/connect4Menu/model/player/Player.java
@@ -1,4 +1,4 @@
-package connect4Menu.model;
+package connect4Menu.model.player;
 
 public class Player {
 
@@ -8,12 +8,12 @@ public class Player {
 	}
 
 	private String name;
-	private PlayerColors color;
+	private IPlayerColors color;
 	private boolean isWinner = false;
 	private boolean isTurn = false;
 	private int playerNum;
 
-	public Player(String name, PlayerColors color, int playerNum) {
+	public Player(String name, IPlayerColors color, int playerNum) {
 		this.name = name;
 		this.color = color;
 		this.playerNum = playerNum;
@@ -23,11 +23,11 @@ public class Player {
 		return name;
 	}
 
-	public PlayerColors getColor() {
+	public IPlayerColors getColor() {
 		return color;
 	}
 
-	public void setColor(PlayerColors color) {
+	public void setColor(IPlayerColors color) {
 		this.color = color;
 	}
