+++ Final_Project/src/main/java/connect4Menu/view/ObserverEndGame.java
@@ -1,6 +1,6 @@
 package connect4Menu.view;
 
-import connect4Menu.model.Player;
+import connect4Menu.model.player.Player;
 
 public interface ObserverEndGame {
